update ubuntu version to add PHP8
~/docker-example-2/Dockerfile 
Changed version of ubuntu to ubuntu:22.10 in order to install php version 8+ 
this php version is needed otherwise Kirby 3.8 will not run. This script as it was will give an error on the 'localhost' kirby page. 

(I am a novice at this, maybe there is a better way to make sure php8 is used. This worked for me)