Added a hint that auth.challenge.email config is also used for password-reset
Took me a moment to find the code creating the actual email for the password-reset and to realize that it also uses the email config from `auth.challenge.email` as the reset is also a challenge from that point of view. This commit drops a hopefully helpful note on that.