fix: Don’t keep HTML responses fresh in browser cache
The `Expires` header affects both the browser and CDN cache. The browser assumes that the response is fresh until the provided date, which lead to stale responses especially during sales. By switching to `Cache-Control: s-maxage`, only shared caches will be affected, which comes close enough to the goal of only affecting the CDN cache